Split Hypergraph class into new notebook
